jimwashkau-site/tools/requirements.txt
openai-whisper
GitPython
sounddevice
numpy
scipy
setuptools
